requirements.txt
eth-ape
ape-etherscan
pandas